refactor(headless-stories): extract shared CSS-modules webpack rule
Source of truth for the CSS-Modules + ?raw webpack wiring lives at
`stories/.storybook/css-modules-webpack.js`. Both the per-package
storybook and the docsite app `require()` it — eliminates the duplicated
~80 LoC of `cssModuleRule` + `patchRules` between the two configs.
